♻️ Refactor: Move queue List to separate folder
