+++ client/src/features/queue/components/Queue.module.scss
@@ -10,16 +10,6 @@
 }
 
 
-
-.queueList {
-  @apply p-2 overflow-y-scroll;
-  height: calc(100% - 64px);
-}
-
-.shadow {
-  @apply h-4 w-full absolute -translate-y-4 bg-transparent shadow-xl shadow-black z-10;
-}
-
 // Card
 .libraryCard {
   @apply p-2 flex items-center gap-3 rounded-md  cursor-pointer;
