+++ client/src/features/queue/components/list/QueueList.module.scss
@@ -1,0 +1,8 @@
+.queueList {
+  @apply p-2 overflow-y-scroll;
+  height: calc(100% - 64px);
+}
+
+.shadow {
+  @apply h-4 w-full absolute -translate-y-4 bg-transparent shadow-xl shadow-black z-10;
+}
